extend api helpers and claim
